// pages/src/components/FindingDetailView.module.css.d.ts
declare const styles: {
  container: string;
  header: string;
  headerTop: string;
  headerContent: string;
  code: string;
  category: string;
  closeButton: string;
  message: string;
  body: string;
  recommendations: string;
  recommendationsTitle: string;
  recommendationsList: string;
  recommendationItem: string;
};
export default styles;